Refactor LRU Cache.
